+++ coding/tree_sitter/typescript_test.go
@@ -2,8 +2,12 @@ package tree_sitter
 
 import (
 	"os"
-	"sidekick/utils"
+	"strings"
 	"testing"
+
+	"sidekick/utils"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func TestGetFileSignaturesStringTypescript(t *testing.T) {
@@ -225,6 +229,267 @@ function add(a: number, b: number): number {
 
 			if symbolsString != test.expected {
 				t.Errorf("Got %s, expected %s", symbolsString, test.expected)
+			}
+		})
+	}
+}
+
+func TestGetSymbolDefinitionTypescript(t *testing.T) {
+	testCases := []struct {
+		name               string
+		symbolName         string
+		code               string
+		expectedDefinition string
+		expectedError      string
+	}{
+		{
+			name:          "empty code",
+			symbolName:    "TestVar",
+			code:          "",
+			expectedError: `symbol not found: TestVar`,
+		},
+		{
+			name:       "function definition",
+			symbolName: "TestFunc",
+			code: `function TestFunc() {
+	console.log("Hello, world!");
+}`,
+			expectedDefinition: `function TestFunc() {
+	console.log("Hello, world!");
+}`,
+		},
+		{
+			name:       "class definition",
+			symbolName: "TestClass",
+			code: `class TestClass {
+	constructor(name, age) {
+		this.name = name;
+		this.age = age;
+	}
+}`,
+			expectedDefinition: `class TestClass {
+	constructor(name, age) {
+		this.name = name;
+		this.age = age;
+	}
+}`,
+		},
+		// TODO bring back once this is fixed: https://github.com/tree-sitter/tree-sitter/issues/2799#issue-2016383906
+		//		{
+		//			name:       "commented function definition",
+		//			symbolName: "TestFunc",
+		//			code: `// TestFunc is a test function.
+		//function TestFunc() {
+		//	console.log("Hello, world!");
+		//}`,
+		//			expectedDefinition: `// TestFunc is a test function.
+		//function TestFunc() {
+		//	console.log("Hello, world!");
+		//}`,
+		//		},
+		//		{
+		//			name:       "commented class definition",
+		//			symbolName: "TestClass",
+		//			code: `// TestClass is a test class.
+		//class TestClass {
+		//	constructor(name, age) {
+		//		this.name = name;
+		//		this.age = age;
+		//	}
+		//}`,
+		//			expectedDefinition: `// TestClass is a test class.
+		//class TestClass {
+		//	constructor(name, age) {
+		//		this.name = name;
+		//		this.age = age;
+		//	}
+		//}`,
+		//		},
+		//		{
+		//			name:       "commented const definition",
+		//			symbolName: "TestConst",
+		//			code: `// TestConst is a test const.
+		//const TestConst = "test";`,
+		//			expectedDefinition: `// TestConst is a test const.
+		//const TestConst = "test";`,
+		//		},
+		{
+			name:               "const definition",
+			symbolName:         "TestConst",
+			code:               `const TestConst = "test";`,
+			expectedDefinition: `const TestConst = "test";`,
+		},
+		{
+			name:               "var definition",
+			symbolName:         "TestVar",
+			code:               `var TestVar = "test";`,
+			expectedDefinition: `var TestVar = "test";`,
+		},
+		{
+			name:          "symbol not found",
+			symbolName:    "NonExistentSymbol",
+			code:          `var TestVar = "test";`,
+			expectedError: `symbol not found: NonExistentSymbol`,
+		},
+		// we include the entire interface definition as part of the function definition
+		{
+			name:       "interface method definition",
+			symbolName: "TestMethod",
+			code: `interface TestInterface {
+	TestMethod();
+	TestMethod2();
+}`,
+			expectedDefinition: `interface TestInterface {
+	TestMethod();
+	TestMethod2();
+}`,
+		},
+		{
+			name:       "call expression",
+			symbolName: "someFunction",
+			code: `somethingElse();
+
+const x = someFunction({
+	foo: 'bar'
+});`,
+			expectedDefinition: `const x = someFunction({
+	foo: 'bar'
+});`,
+		},
+		{
+			name:       "exported call expression with extra comment",
+			symbolName: "someFunction",
+			code: `somethingElse();
+
+export const x = someFunction({
+	foo: 'bar'
+}); // testing`,
+			expectedDefinition: `export const x = someFunction({
+	foo: 'bar'
+}); // testing`,
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			filePath, err := utils.WriteTestTempFile(t, "ts", tc.code)
+			if err != nil {
+				t.Fatalf("Failed to write temp file: %v", err)
+			}
+			defer os.Remove(filePath)
+
+			definition, err := GetSymbolDefinitionsString(filePath, tc.symbolName, 0)
+			if err != nil {
+				if tc.expectedError == "" {
+					t.Fatalf("Unexpected error: %v", err)
+				} else if !strings.Contains(err.Error(), tc.expectedError) {
+					t.Fatalf("Expected error: %s, got: %v", tc.expectedError, err)
+				}
+			}
+
+			if strings.TrimSuffix(definition, "\n") != strings.TrimSuffix(tc.expectedDefinition, "\n") {
+				t.Errorf("Expected definition:\n%s\nGot:\n%s", utils.PanicJSON(tc.expectedDefinition), utils.PanicJSON(definition))
+			}
+		})
+	}
+}
+
+func TestGetFileHeadersStringTypescript(t *testing.T) {
+	testCases := []struct {
+		name     string
+		code     string
+		expected string
+	}{
+		{
+			name:     "empty",
+			code:     "",
+			expected: "",
+		},
+		{
+			name:     "no imports",
+			code:     "const foo = 'bar';",
+			expected: "",
+		},
+		{
+			name:     "single import",
+			code:     "import { foo } from 'bar';",
+			expected: "import { foo } from 'bar';\n",
+		},
+		{
+			name:     "single import with whitespace",
+			code:     " import { foo } from 'bar';",
+			expected: " import { foo } from 'bar';\n",
+		},
+		{
+			name:     "multiple imports",
+			code:     "import { foo, foo2 } from 'bar';\nimport { baz } from 'qux';",
+			expected: "import { foo, foo2 } from 'bar';\nimport { baz } from 'qux';\n",
+		},
+		{
+			name:     "import with alias",
+			code:     "import { foo as f } from 'bar';",
+			expected: "import { foo as f } from 'bar';\n",
+		},
+		{
+			name:     "import with default",
+			code:     "import foo from 'bar';",
+			expected: "import foo from 'bar';\n",
+		},
+		{
+			name:     "import with namespace",
+			code:     "import * as foo from 'bar';",
+			expected: "import * as foo from 'bar';\n",
+		},
+		{
+			name:     "import with side effects",
+			code:     "import 'bar';",
+			expected: "import 'bar';\n",
+		},
+		{
+			name:     "import with type only",
+			code:     "import type { foo } from 'bar';",
+			expected: "import type { foo } from 'bar';\n",
+		},
+		{
+			name:     "import with type and side effects",
+			code:     "import type 'bar';",
+			expected: "import type 'bar';\n",
+		},
+		{
+			name:     "import with type and default",
+			code:     "import type foo from 'bar';",
+			expected: "import type foo from 'bar';\n",
+		},
+		{
+			name:     "import with type and namespace",
+			code:     "import type * as foo from 'bar';",
+			expected: "import type * as foo from 'bar';\n",
+		},
+		{
+			name:     "nested imports",
+			code:     "function x() {\n    import { foo } from 'bar';\n    import { baz } from 'qux';\n}",
+			expected: "",
+		},
+	}
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			// Create a temporary file with the test case code
+			tmpfile, err := os.CreateTemp("", "test*.ts")
+			if err != nil {
+				t.Fatalf("Failed to create temp file: %v", err)
+			}
+			defer os.Remove(tmpfile.Name())
+			if _, err := tmpfile.Write([]byte(tc.code)); err != nil {
+				t.Fatalf("Failed to write to temp file: %v", err)
+			}
+			if err := tmpfile.Close(); err != nil {
+				t.Fatalf("Failed to close temp file: %v", err)
+			}
+			result, err := GetFileHeadersString(tmpfile.Name(), 0)
+			assert.Nil(t, err)
+			// Check the result
+			if result != tc.expected {
+				t.Errorf("GetFileHeadersString returned incorrect result. Expected:\n%s\nGot:\n%s", utils.PanicJSON(tc.expected), utils.PanicJSON(result))
 			}
 		})
 	}
